+++ src/test/java/lotto/LottoIssueRequestTest.java
@@ -1,5 +1,6 @@
 package lotto;
 
+import lotto.application.LottoIssueRequest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
