+++ src/test/java/lotto/LottoResultsTest.java
@@ -1,5 +1,7 @@
 package lotto;
 
+import lotto.domain.LottoResults;
+import lotto.domain.Rank;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
